# target/classes/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Account Management API Documentation
  description: API documentation for the Account Management System
servers:
  - url: http://127.0.0.1:8080
    description: Local server
paths:
...
  /transactions:
    get:
      summary: Retrieve all transactions
      operationId: getAllTransactions
      description: Retrieve all transactions
      tags:
        - Transaction
      responses:
        200:
          description: A list of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountTransaction'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /transactions/{id}:
    get:
      summary: Retrieve a transaction by ID
      operationId: getTransactionById
      description: Retrieve a transaction by ID
      tags:
        - Transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountTransaction'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /transactions/account/{accountId}:
    get:
      summary: Retrieve transactions by account ID
      operationId: getTransactionsByAccountId
      description: Retrieve transactions by account ID
      tags:
        - Transaction
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Transactions for the account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountTransaction'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
...
  /accountTypes:
    get:
      summary: Retrieve all account types
      operationId: getAllAccountTypes
      description: Retrieve all account types
      tags:
        - Type
      responses:
        200:
          description: A list of account types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountType'
        500:
          $ref: '#/components/responses/ErrorResponse'
  /accountTypes/{id}:
    get:
      summary: Retrieve an account type by ID
      operationId: getAccountTypeById
      description: Retrieve an account type by ID
      tags:
        - Type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Account type details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountType'
        404:
          $ref: '#/components/responses/ErrorResponse'
        500:
          $ref: '#/components/responses/ErrorResponse'
components:
  schemas:
...
    AccountTransaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        transactionType:
          type: string
        amount:
          type: number
          format: double
        transactionDate:
          type: string
          format: date-time
        description:
          type: string
...
    AccountType:
      type: object
      properties:
        id:
          type: integer
          format: int64
        typeName:
          type: string
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: string